Fix allocation bug for index data

+++ OpenGLSandbox/OpenGLSandbox/src/Renderer/IndexBuffer.cpp
@@ -91,7 +91,7 @@ void IndexBuffer::loadFromFile(std::string fileName)
 
     inputFile.close();
 
-    glBufferData(m_targetType, sizeof(GLfloat) * m_indices.size(), m_indices.data(), m_usageType);
+    glBufferData(m_targetType, sizeof(GLshort) * m_indices.size(), m_indices.data(), m_usageType);
 }
 
 void IndexBuffer::push(GLshort p1, GLshort p2, GLshort p3)
@@ -103,7 +103,7 @@ void IndexBuffer::push(GLshort p1, GLshort p2, GLshort p3)
 
 void IndexBuffer::uploadData()
 {
-    glBufferData(m_targetType, sizeof(GLfloat) * m_indices.size(), m_indices.data(), m_usageType);
+    glBufferData(m_targetType, sizeof(GLshort) * m_indices.size(), m_indices.data(), m_usageType);
 }
 
 GLsizei IndexBuffer::getSize() const
